+++ test/tiles_writer.cpp
@@ -1,12 +1,11 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
+#include "tiles_test_util.h"
 #include "util.h"
 
 #include <algorithm>
 #include <cstdint>
-#include <cstdlib>
-#include <ctime>
 #include <filesystem>
 #include <fstream>
 #include <functional>
@@ -62,13 +61,8 @@ static void overwrite_file(
 
 int main()
 {
-  const auto seed = std::time(nullptr);
-  std::srand((unsigned)seed);
-  std::cout << "seed=" << seed << '\n';
-
-  const fs::path base = fs::temp_directory_path() /
-    ("merklecpp_tiles_writer_" + std::to_string((unsigned long long)seed) +
-     "_" + std::to_string(std::rand()));
+  const TemporaryDirectory temporary_directory("merklecpp_tiles_writer");
+  const fs::path& base = temporary_directory.path();
 
   try
   {
@@ -316,22 +310,15 @@ int main()
     }
 
     std::cout << "tiles_writer: OK" << '\n';
-
-    std::error_code ec;
-    fs::remove_all(base, ec);
   }
   catch (std::exception& ex)
   {
     std::cout << "Error: " << ex.what() << '\n';
-    std::error_code ec;
-    fs::remove_all(base, ec);
     return 1;
   }
   catch (...)
   {
     std::cout << "Error" << '\n';
-    std::error_code ec;
-    fs::remove_all(base, ec);
     return 1;
   }
